Progres splitting L2 prod tests into files
