+++ test/optimism/utils/rpc.js
@@ -1,23 +1,3 @@
-const fs = require('fs');
-const path = require('path');
-const ethers = require('ethers');
-const { getSource, getTarget } = require('../..');
-
-function connectContract({ contract, source = contract, provider, useOvm = false }) {
-	const params = {
-		path,
-		fs,
-		network: 'local',
-		useOvm,
-	};
-
-	return new ethers.Contract(
-		getTarget({ ...params, contract }).address,
-		getSource({ ...params, contract: source }).abi,
-		provider
-	);
-}
-
 async function wait(seconds) {
 	return new Promise(resolve => {
 		setTimeout(() => {
@@ -44,9 +24,8 @@ async function restoreSnapshot({ id, provider }) {
 }
 
 module.exports = {
-	connectContract,
-	wait,
-	fastForward,
 	takeSnapshot,
 	restoreSnapshot,
+	fastForward,
+	wait,
 };
